requirements.txt
python-telegram-bot>=21,<22
playwright>=1.30,<2.0
markdown-it-py>=3,<4
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2,<3
python-dotenv>=1,<2
aiofiles>=23,<24
